requirements.txt
python-dotenv==0.14.0
python-telegram-bot==13.0
pymongo==3.11.1
python-twitter==3.5